% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/recuperation_donnees_SIM2_quotidien_listeStations.R
\name{recuperation_donnees_SIM2_quotidien_listeStations}
\alias{recuperation_donnees_SIM2_quotidien_listeStations}
\title{Récupération des données SIM2 Météo France quotidiennes d'une liste de points de grille}
\usage{
recuperation_donnees_SIM2_quotidien_listeStations(liste_LAMBXLAMBY_meteoFrance)
}
\arguments{
\item{liste_LAMBXLAMBY_meteoFrance}{une chaine de caractère où LAMBX et LAMBY sont séparées par un espace}
}
\value{
les données météorologiques des points de grille identifiés, depuis le 1er du mois précédent
}
\description{
Récupération des données SIM2 Météo France quotidiennes d'une liste de points de grille
}
\examples{
# Identifiant des points de grille situés dans la vallée du Grésivaudan : LAMBX LAMBY
data_list <- recuperation_donnees_SIM2_quotidien_listeStations(c("760 24170", "8680 20250"))
}
