% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recuperation_donnees_SIM2_quotidiennes.R
\name{recuperation_donnees_SIM2_quotidiennes}
\alias{recuperation_donnees_SIM2_quotidiennes}
\title{Récupération des données SIM2 Météo France quotidiennes d'un point de grille}
\usage{
recuperation_donnees_SIM2_quotidiennes(LAMBXLAMBY_meteoFrance)
}
\arguments{
\item{LAMBXLAMBY_meteoFrance}{une chaine de caractère où LAMBX et LAMBY sont séparées par un espace}
}
\value{
les données météorologiques du point de grille identifié, depuis le 1er du mois précédent
}
\description{
Récupération des données SIM2 Météo France quotidiennes d'un point de grille
}
\examples{
# Identifiant du point de grille situé sur Grenoble : LAMBX LAMBY
recuperation_donnees_SIM2_quotidiennes("8680 20250")
}
