% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recuperation_points_de_grille_SAFRAN.R
\name{recuperation_points_de_grille_SAFRAN}
\alias{recuperation_points_de_grille_SAFRAN}
\title{Récupération des identifiants des points de grille SAFRAN}
\usage{
recuperation_points_de_grille_SAFRAN()
}
\value{
la localisation de tous les points de grille disponibles
}
\description{
Récupération des identifiants des points de grille SAFRAN
}
\examples{
# Récupération de tous les points de grille disponibles
points_de_grille <- recuperation_points_de_grille_SAFRAN()
}
